First attempt at using a mixin class
